Refaractoring of code completed
