+++ lib/instagram_ui_clone/common/route_builders.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+Route slideUpRoute({Widget routeName}) {
+  return PageRouteBuilder(
+    pageBuilder: (context, animation, secondaryAnimation) => routeName,
+    transitionsBuilder: (context, animation, secondaryAnimation, child) {
+      var begin = Offset(0.0, 1.0);
+      var end = Offset.zero;
+      var curve = Curves.ease;
+
+      var tween = Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
+
+      return SlideTransition(
+        position: animation.drive(tween),
+        child: child,
+      );
+    },
+  );
+}
