+++ lib/instagram_ui_clone/screens/add_post_camera.dart
@@ -1,21 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_example/instagram_ui_clone/common/utils.dart';
 import 'package:flutter_example/instagram_ui_clone/widgets/add_post_app_bar.dart';
-import 'package:flutter_icons/flutter_icons.dart';
+import 'package:flutter_example/instagram_ui_clone/widgets/add_post_camera_button.dart';
 
 class AddPostCamera extends StatelessWidget {
   const AddPostCamera({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    Size _screen = MediaQuery.of(context).size;
     return Scaffold(
       appBar: AddPostAppBar(
         screenName: 'camera',
         height: 45,
       ),
       body: Container(
-        height: MediaQuery.of(context).size.height,
-        width: MediaQuery.of(context).size.width,
+        height: _screen.height,
+        width: _screen.width,
         child: Column(
           children: <Widget>[
             Container(
@@ -27,12 +28,12 @@ class AddPostCamera extends StatelessWidget {
                     image: NetworkImage(Utils.listOfImageUrl.elementAt(10)),
                     fit: BoxFit.cover,
                   ),
-                  _cameraOptions(),
+                  cameraOptions(),
                 ],
               ),
             ),
             Container(
-              height: MediaQuery.of(context).size.height - 410 - 45 - 45,
+              height: _screen.height - 410 - 45 - 45,
               color: Theme.of(context).primaryColorDark,
               child: Stack(
                 children: <Widget>[
@@ -61,32 +62,6 @@ class AddPostCamera extends StatelessWidget {
                 ],
               ),
             ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _cameraOptions() {
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: Container(
-        height: 60,
-        padding: EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: <Widget>[
-            Icon(
-              Feather.rotate_ccw,
-              color: Colors.white,
-              size: 30,
-            ),
-            Icon(
-              Ionicons.ios_flash,
-              color: Colors.white,
-              size: 30,
-            )
           ],
         ),
       ),
