+++ lib/instagram_ui_clone/screens/home_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_example/instagram_ui_clone/common/route_builders.dart';
 import 'package:flutter_example/instagram_ui_clone/data/models/post.dart';
 import 'package:flutter_example/instagram_ui_clone/data/models/story.dart';
 import 'package:flutter_example/instagram_ui_clone/data/repositories/feed_repo.dart';
@@ -6,8 +7,7 @@ import 'package:flutter_example/instagram_ui_clone/screens/activity_screen.dart'
 import 'package:flutter_example/instagram_ui_clone/screens/add_post_screen.dart';
 import 'package:flutter_example/instagram_ui_clone/screens/explore_screen.dart';
 import 'package:flutter_example/instagram_ui_clone/screens/feed_screen.dart';
-import 'package:flutter_example/instagram_ui_clone/screens/profile_screen_sliver.dart';
-import 'package:flutter_example/instagram_ui_clone/screens/profile_screen_v1.dart';
+import 'package:flutter_example/instagram_ui_clone/screens/profile_screen.dart';
 import 'package:flutter_statusbar_manager/flutter_statusbar_manager.dart';
 
 class HomeScreen extends StatefulWidget {
@@ -39,7 +39,7 @@ class _HomeScreenState extends State<HomeScreen> {
       ExploreScreen(),
       AddPostScreen(),
       ActivityScreen(),
-      ProfileScreenV1(),
+      ProfileScreen(),
     ];
 
     return Scaffold(
@@ -51,8 +51,14 @@ class _HomeScreenState extends State<HomeScreen> {
         currentIndex: pageIndex,
         onTap: (index) {
           if (index == 2) {
-            FlutterStatusbarManager.setHidden(true)
-                .then((_) => Navigator.push(context, _slideUpRoute()));
+            FlutterStatusbarManager.setHidden(true).then(
+              (_) => Navigator.push(
+                context,
+                slideUpRoute(
+                  routeName: AddPostScreen(),
+                ),
+              ),
+            );
           } else {
             setState(() {
               pageIndex = index;
@@ -84,25 +90,6 @@ class _HomeScreenState extends State<HomeScreen> {
           ),
         ],
       ),
-    );
-  }
-
-  Route _slideUpRoute() {
-    return PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => AddPostScreen(),
-      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-        var begin = Offset(0.0, 1.0);
-        var end = Offset.zero;
-        var curve = Curves.ease;
-
-        var tween =
-            Tween(begin: begin, end: end).chain(CurveTween(curve: curve));
-
-        return SlideTransition(
-          position: animation.drive(tween),
-          child: child,
-        );
-      },
     );
   }
 }
